+++ packages/core/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### 1.6.0-alpha.1 - 2021-02-01
+
+#### 🐞 Fixes
+
+- add useGraphQLToolsUnorderedDirectives option to directive plugin ([6517dd4](https://github.com/hayes/giraphql/commit/6517dd4))
+
+**Note:** Version bump only for package @giraphql/core
+
+
+
+
+
+## 1.6.0-alpha.0 - 2021-02-01
+
+#### 🚀 Updates
+
+- add directives plugin ([ff76fe7](https://github.com/hayes/giraphql/commit/ff76fe7))
+
+**Note:** Version bump only for package @giraphql/core
+
+
+
+
+
 ### 1.5.1 - 2021-01-26
 
 #### 🐞 Fixes